requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==2.0.12
Django==3.2.22
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
idna==3.4
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.26.0
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==1.26.17
